+++ lib/jss/api_object/computer.rb
@@ -44,7 +44,12 @@ module JSS
   # * barcodes 1 and 2
   # * ip_address
   # * udid
-  # * mac_addresses
+  # * mac_address & alt_mac_address
+  # * serial_number
+  #   Note: Even tho the webUI doesn't allow editing the serial_number,
+  #   the API does, and it can be useful for dealing with duplicates
+  #   that arise when a logic-board swap causes a new computer record.
+  #   to be created.
   # * location data via the Locatable module
   # * purchasing data via the Purchasable module
   # * Extension Attribute values via the Extendable module
@@ -1126,10 +1131,35 @@ module JSS
     #
     def ip_address=(new_val)
       return nil if @ip_address == new_val
-      new_val.strip!
-      # this raises an error if its an invalid IP address
-      IPAddr.new new_val
-      @ip_address = new_val
+      @ip_address = new_val.empty? ? new_val : JSS::Validate.ip_address(new_val)
+      @need_to_update = true
+    end
+
+    #
+    def mac_address=(new_val)
+      return nil if new_val == @mac_address
+      @mac_address =  new_val.empty? ? new_val : JSS::Validate.mac_address(new_val)
+      @need_to_update = true
+    end
+
+    #
+    def alt_mac_address=(new_val)
+      return nil if new_val == @alt_mac_address
+      @alt_mac_address = new_val.empty? ? new_val : JSS::Validate.mac_address(new_val)
+      @need_to_update = true
+    end
+
+    #
+    def serial_number=(new_val)
+      return nil if new_val == @serial_number
+      @serial_number =  new_val.empty? ? new_val : JSS::Validate.unique_identifier(JSS::Computer, :serial_number, new_val)
+      @need_to_update = true
+    end
+
+    #
+    def udid=(new_val)
+      return nil if new_val == @udid
+      @udid = new_val.empty? ? new_val : JSS::Validate.unique_identifier(JSS::Computer, :udid, new_val)
       @need_to_update = true
     end
 
@@ -1268,6 +1298,7 @@ module JSS
       general.add_element('ip_address').text = @ip_address
       general.add_element('mac_address').text = @mac_address
       general.add_element('udid').text = @udid
+      general.add_element('serial_number').text = @serial_number
 
       rmgmt = general.add_element('remote_management')
       rmgmt.add_element('managed').text = @managed